se creo la interface abstractInterface y se le heredo a cada interface de cada facade

+++ src/main/java/Facades/AbstractFacade.java
@@ -22,16 +22,30 @@ public abstract class AbstractFacade<T> {
 
     protected abstract EntityManager getEntityManager();
 
-    public void create(T entity) {
+        public T create(T entity) {
+        if(entity != null && getEntityManager() !=null){
         getEntityManager().persist(entity);
+        return entity;
+        }
+        else{return null;}
+        
     }
 
-    public void edit(T entity) {
+    public T edit(T entity) {
+        if(entity != null && getEntityManager() != null){
         getEntityManager().merge(entity);
+        return entity;
+        }
+        else{return null;}
     }
 
-    public void remove(T entity) {
+    public T remove(T entity) {
+        if(entity != null && getEntityManager() != null){
         getEntityManager().remove(getEntityManager().merge(entity));
+        return entity;
+        }
+        else{return null;}
+        
     }
 
     public T find(Object id) {
